Add files via upload
This first code will take the file created by ORDEM and focus on the Butterlfy flux. Data are imported and stored in a list, and the user can choose the size of interest. Therefore, the list will contain all the flux of debris between [val_min and val_max], depending on the angle (from the satellite frame).